+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/SampleSingleCell.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public interface SampleSingleCell extends SampleTissueProcessing {
 
-  public static final String SAMPLE_CLASS_NAME = "Single Cell";
+  public static final String SUBCATEGORY_NAME = "Single Cell";
 
   public BigDecimal getInitialCellConcentration();